mod organization and documentation
