+++ card/lib/card/mod/dirs.rb
@@ -1,4 +1,13 @@
 class Card
+
+  def config
+    Cardio.config
+  end
+
+  def paths
+    Cardio.paths
+  end
+
   module Mod
     # Dirs objects are used to manage the load paths for card mods.
     # Mods can be loaded as gems and by using directories with mod subdirectories.
