--- card/lib/card/mod/loader/module_loader/load_strategy/eval.rb
@@ -1,20 +1,0 @@
-class Card
-  module Mod
-    module Loader
-      class ModuleLoader
-        class LoadStrategy
-          # Put everything for the module definition in one string and the evaluate
-          # it immediately with ruby's eval method.
-          class Eval < Card::Mod::Loader::ModuleLoader::LoadStrategy
-            def load_modules
-              each_file do |abs_path, module_names|
-                template = @module_template.new module_names, abs_path
-                template.build
-              end
-            end
-          end
-        end
-      end
-    end
-  end
-end
