+++ card/mod/carrierwave/set/abstract/attachment/paths.rb
@@ -22,7 +22,7 @@ end
 
 def mod_dir new_mod=nil
   mod_name = new_mod || mod
-  dir = Card::Mod::Loader.mod_dirs.path(mod_name) || (mod_name.to_sym == :test && "test")
+  dir = Mod.dirs.path(mod_name) || (mod_name.to_sym == :test && "test")
 
   raise Error, "can't find mod \"#{mod_name}\"" unless dir
   dir
