+++ card/mod/carrierwave/spec/set/type/file_spec.rb
@@ -218,11 +218,11 @@ describe Card::Set::Type::File do
           # File.open(File.join(file_dir, "test_mod.txt"),"w") do |f|
           #   f.puts "test"
           # end
-          Card::Mod::Loader.mod_dirs.mod "test_mod"
+          Card::Mod.dirs.mod "test_mod"
         end
         after do
           FileUtils.rm_rf mod_path
-          Card::Mod::Loader.mod_dirs.mods.delete "test_mod"
+          Card::Mod.dirs.mods.delete "test_mod"
         end
         subject do
           create_file_card :coded, test_file,
@@ -348,11 +348,11 @@ describe Card::Set::Type::File do
     context "if storage type is coded" do
       before do
         FileUtils.mkdir_p mod_path
-        Card::Mod::Loader.mod_dirs.mod "test_mod"
+        Card::Mod.dirs.mod "test_mod"
       end
       after do
         FileUtils.rm_rf mod_path
-        Card::Mod::Loader.mod_dirs.mods.delete "test_mod"
+        Card::Mod.dirs.mods.delete "test_mod"
       end
 
       subject do
