+++ card/spec/lib/card/mod/loader/module_loader/pattern_loader_spec.rb
@@ -8,7 +8,7 @@ RSpec.describe Card::Mod::Loader::ModuleLoader::PatternLoader do
   end
 
   it 'initializes the load strategy with the right module type and module template' do
-    expect(Card::Mod::Loader::ModuleLoader::LoadStrategy::Eval)
+    expect(Card::Mod::LoadStrategy::Eval)
       .to receive(:new).with(
         mod_dirs, :set_pattern,
         Card::Mod::Loader::ModuleTemplate::PatternModule)
